set up authentication and authorization using JWT

+++ controllers/api/v1/posts_api.js
@@ -32,19 +32,21 @@ module.exports.destroy = async function(req,res){
     try{
         let post = await Post.findById(req.params.id);
 
-        // if(post.user == req.user.id){
-        post.remove();
-        await Comment.deleteMany({post: req.params.id});         
+        if(post.user == req.user.id){
+            
+            post.remove();
+            await Comment.deleteMany({post: req.params.id});         
 
-        return res.json(200, {
-            message: "Post and associated comments deleted successfully"
-        });
+            return res.json(200, {
+                message: "Post and associated comments deleted successfully"
+            });
 
-        // }else{
+        }else{
 
-        //     req.flash('error', 'You cannot delete this post!');
-        //     return res.redirect('back');
-        // }
+            return res.json(401, {
+                message: 'You cannot delete this post!'
+            })
+        }
 
     }catch(err){
